+++ download-utilities.sh
@@ -1,5 +1,5 @@
 #!/bin/bash
-# download-utilities v2.8
+# download-utilities v2.9
 # Created by Sam Gleske
 # Fri 19 May 2023 06:01:53 PM EDT
 # MIT Licensed; https://github.com/samrocketman/yml-install-files